Create issue_template.md
This is a template for raising issues in our repo. From this point forward, all issues should follow this template to make it easier for us to understand the issue without needing to look into the code ourselves. Additionally, please indicate whether or not the issue has been unit tested or not since this could be a way for us to add more tests that we might have missed during solo development.