Dashboard
# Summary
updating the skeleton code and further laying out what needs to be done. nothing complex.

# Description
this is bare bones skeleton code. until we finalize the way we are linking pages that feature will be left unfinished. what is on the page are div containers and buttons for the compare students and alerts.

# Documentation
variables used for the page will be the following:
    Variables: Class_Name, Section_Name, Display_Name
    Buttons: Menu, Student Comparison, Alerts, n boxes with Class_Names and Section_Names
    Links: Student Comparison Tool, Alerts, Each box
    Header takes Display_Name

documentation on how the page will function is found in the discord - additional documentation will be made at a later time. the wireframe is from Jacobs wireframe in the discord.

# Unit Tests
no unit tests were made are there is no functionality currently to test

# PR Reviews
front end should review and see if any changes or additions need to be made to the structure of the code.

for what needs to be done:
1. linking of the pages to the compare students and alerts pages
2. creating and linking to the sections and their assignments
3. creation of the banner and header for the welcome text
4. further stylization and completion

![image_2025-03-24_201331187](https://github.com/user-attachments/assets/997355b9-ca34-41ae-92d9-71a04606fcdb)

once linking and querey are added, tests will be made to ensure that the query is correct and that the changing of the pages is successful.