Added new issue templates
These will be the new templates that will appear when an issue is submitted in the PRISM repo. The old issue template is still present, but there is now a template for features that should be utilized whenever someone plans on working on a new feature not present in PRISM yet.