Update student_comparison
This is my current state of student comparison. It uses a lot of useState variables and if statements in order to get conditional rendering. Also, it contains the current api calls AS I KNOW THEM. The only parts left are implementing getting the correct classes based on professor, getting conditional disables on the lower dropdowns, and testing.